// src/common/enum/env.enum.ts
export enum ENV_KEY {
  NODE_ENV = 'NODE_ENV',
  APP_PORT = 'APP_PORT',

  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_USER = 'DB_USER',
  DB_PASS = 'DB_PASS',
  DB_NAME = 'DB_NAME',

  DB_SYNCHRONIZE = 'app.synchronize',
}

export enum NODE_ENV {
  production = 'production',
  development = 'development',
}